fix(backend): resolve IDOR, atomic balance sync, and update schemas for cents

+++ budget-bee-backend/src/controllers/incomeSources.ts
@@ -6,7 +6,7 @@ import { logger } from '../utils/logger.js';
 
 const incomeSourceSchema = z.object({
     name: z.string().min(1),
-    amount: z.number().positive(),
+    amount: z.number().int().positive(),
     frequency: z.enum(['Monthly', 'Weekly', 'Bi-weekly', 'Annually']),
 });
